+++ js/apps/system/_admin/aardvark/APP/react/src/views/views/arangoSearchView/linksEditor/AnalyzersDropdown.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { MultiSelectControl } from "../../../../components/form/MultiSelectControl";
+import { useEditViewContext } from "../../editView/EditViewContext";
+
+export const AnalyzersDropdown = () => {
+  const { currentLink } = useEditViewContext();
+  return (
+    <MultiSelectControl
+      name={`links.${currentLink}.analyzers`}
+      label="Analyzers"
+    />
+  );
+};
