+++ js/apps/system/_admin/aardvark/APP/react/src/views/views/editView/EditViewWrap.tsx
@@ -14,6 +14,7 @@ import { SearchViewsCustomStyleReset } from "../SearchViewsCustomStyleReset";
 
 export const EditViewWrap = ({ name }: { name: string }) => {
   const { view, isLoading } = useFetchViewProperties(name);
+  console.log({ view, isLoading });
   useDisableNavBar();
   useGlobalStyleReset();
   if (isLoading) {
